resume-shortlisting-system/ml-service/requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
nltk>=3.8.1
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
openai>=0.28.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=0.8.11
requests>=2.31.0
